Refactor server
Simplifies the websocket loop logic from "wait for two messages, first
one tells the current open route (to be watched), second one tells the
route to be sent" to "on every message, switch to its route and remember
that it is being watched".

Removes routeFromPathInfo (that would always strip the first character
from the websocket message) and instead assumes the js shim strips the
base path completely.

This is in preparation to the server customization feature.

Also simplify parts of the runServerWithWebSocketHotReload function.