Simplify the websocket loop logic
Instead of waiting for two messages in each refresh, watch the last
route that was asked by the client.

This works because there is no way for the client to move to a different
page without either asking the webserver, or causing the client to
disconnect. Upon reconnection the client asks for the new route again.